requirements.txt
boto3 == 1.4.4
coveralls == 1.1
mock == 1.3.0
requests == 2.2.0
scipy == 0.19.0
python-slugify == 1.2.4
statistics == 1.0.3.5
statsmodels == 0.6.1
